requirements.txt
Django>=5.2.8,<6.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0
